followup refactors
