+++ packages/realm-server/prerender/utils.ts
@@ -169,10 +169,11 @@ function buildInvalidRenderResponseError(
   try {
     let pathname = new URL(page.url()).pathname;
     let match = /\/render\/([^/]+)\//.exec(pathname);
-    id = match?.[1] ?? null;
+    id = match?.[1] ? decodeURIComponent(match[1]) : null;
   } catch {
     id = null;
   }
+  let deps = fallbackRenderDeps(id);
   return {
     type: 'instance-error',
     error: {
@@ -181,9 +182,23 @@ function buildInvalidRenderResponseError(
       title: options?.title ?? 'Invalid render response',
       message,
       additionalErrors: null,
+      deps,
     },
     ...(options?.evict ? { evict: true } : {}),
   };
+}
+
+function fallbackRenderDeps(id: string | null): string[] {
+  if (!id) {
+    return [];
+  }
+  let deps = new Set<string>([id]);
+  if (id.endsWith('.json')) {
+    deps.add(id.replace(/\.json$/, ''));
+  } else {
+    deps.add(`${id}.json`);
+  }
+  return [...deps];
 }
 
 export function buildInvalidModuleResponseError(
@@ -718,17 +733,28 @@ export async function captureResult(
             })
           | undefined;
         if (!firstChild) {
+          let cardId = resolvedElement.dataset.prerenderId ?? null;
+          let deps = cardId
+            ? Array.from(
+                new Set(
+                  cardId.endsWith('.json')
+                    ? [cardId, cardId.replace(/\.json$/, '')]
+                    : [cardId, `${cardId}.json`],
+                ),
+              )
+            : [];
           return {
             status: 'error',
             value: JSON.stringify({
               type: 'instance-error',
               error: {
-                id: resolvedElement.dataset.prerenderId ?? null,
+                id: cardId,
                 status: 500,
                 title: 'Invalid render response',
                 message:
                   '[data-prerender] has no child element to capture (render produced no root element)',
                 additionalErrors: null,
+                deps,
               },
             }),
             alive,
