+++ packages/runtime-common/index-runner/dependency-normalization.ts
@@ -1,0 +1,61 @@
+import { trimExecutableExtension } from '../index';
+import { canonicalURL } from './dependency-url';
+
+export function isExtensionlessPath(url: URL): boolean {
+  let lastSegment = url.pathname.split('/').pop() ?? '';
+  return !lastSegment.includes('.');
+}
+
+export function normalizeStoredDependency(
+  dep: string,
+  relativeTo: URL,
+): string {
+  return canonicalURL(dep, relativeTo.href);
+}
+
+export function normalizeRelationshipDependency(
+  dep: string,
+  relativeTo: URL,
+  realmURL: URL,
+): string {
+  let canonical = canonicalURL(dep, relativeTo.href);
+  try {
+    let normalized = new URL(canonical);
+    if (
+      normalized.href.startsWith(realmURL.href) &&
+      isExtensionlessPath(normalized)
+    ) {
+      normalized.pathname = `${normalized.pathname}.json`;
+    }
+    return normalized.href;
+  } catch (_err) {
+    return canonical;
+  }
+}
+
+export function canTraverseRelationshipDependency(
+  dep: string,
+  realmURL: URL,
+): boolean {
+  try {
+    let parsed = new URL(dep);
+    if (!parsed.href.startsWith(realmURL.href)) {
+      return false;
+    }
+    return !isExtensionlessPath(parsed);
+  } catch (_err) {
+    return false;
+  }
+}
+
+export function normalizeDependencyForLookup(
+  dep: string,
+  relativeTo: URL,
+): string {
+  let canonical = canonicalURL(dep, relativeTo.href);
+  try {
+    return trimExecutableExtension(new URL(canonical)).href;
+  } catch (_err) {
+    return canonical;
+  }
+}
